tox-python: installing pyenv in the tox home
also there are build args for user and group id. if you need a different
user or group id, you have to use the build args. i know its better to
do this in a entrypoint but i had some problems with permissions for
python cache files. like this everything is under the right user.